sql: infra changes for support atlas:import directive
